requirements.txt
# ===============================================================================
# PROJECT SENTINEL - DEPENDENCIES
# ===============================================================================
# Install with: uv pip install -r requirements.txt
# ===============================================================================

# Core
MetaTrader5>=5.0.45
python-dotenv>=1.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# HTTP & API
requests>=2.28.0

# System Monitoring
psutil>=5.9.0

# Encryption (optional but recommended)
cryptography>=41.0.0

# Database
# (sqlite3 is built-in)
